+++ main.py
@@ -1,15 +1,15 @@
-from pydantic import BaseModel, Field
-from typing import Optional
+from fastapi import FastAPI
+from fastapi.responses import HTMLResponse
+from fastapi.openapi.docs import get_swagger_ui_html
+from router import router
+app = FastAPI()
 
-# Modelo para la respuesta de datos de la tabla grades (con id)
-class Grade(BaseModel):
-    id: int
-    student_id: int
-    subject_id: int
-    grade: float 
-
-# Modelo para la creación de un nuevo registro de calificación (sin id)
-class GradeCreate(BaseModel):
-    student_id: int
-    subject_id: int
-    grade: float = Field(..., ge=1.0, le=5.0, description="Grade must be between 1.0 and 5.0")
+@app.get("/", response_class=HTMLResponse)
+def custom_swagger_ui_html():
+    return get_swagger_ui_html(
+        openapi_url="/openapi.json",
+        title="school",
+        swagger_js_url="https://cdn.jsdelivr.net/npm/swagger-ui-dist@5/swagger-ui-bundle.js",
+        swagger_css_url="https://cdn.jsdelivr.net/npm/swagger-ui-dist@5/swagger-ui.css",
+    )
+app.include_router(router)
